+++ client/js/src/client.ts
@@ -42,7 +42,7 @@ type predict = (
 	event_data?: unknown
 ) => Promise<unknown>;
 
-type client_return = {
+export type client_return = {
 	predict: predict;
 	config: Config;
 	submit: (
@@ -194,6 +194,7 @@ export function api_factory(
 		if (token) {
 			headers.Authorization = `Bearer ${token}`;
 		}
+
 		try {
 			var response = await fetch_implementation(url, {
 				method: "POST",
@@ -758,9 +759,11 @@ export function api_factory(
 						} else if (
 							protocol == "sse_v1" ||
 							protocol == "sse_v2" ||
-							protocol == "sse_v2.1"
+							protocol == "sse_v2.1" ||
+							protocol == "sse_v3"
 						) {
 							// latest API format. v2 introduces sending diffs for intermediate outputs in generative functions, which makes payloads lighter.
+							// v3 only closes the stream when the backend sends the close stream message.
 							fire_event({
 								type: "status",
 								stage: "pending",
@@ -855,7 +858,7 @@ export function api_factory(
 												});
 												if (
 													data &&
-													(protocol === "sse_v2" || protocol === "sse_v2.1")
+													["sse_v2", "sse_v2.1", "sse_v3"].includes(protocol)
 												) {
 													apply_diff_stream(event_id!, data);
 												}
@@ -904,7 +907,9 @@ export function api_factory(
 												fn_index,
 												time: new Date()
 											});
-											close_stream();
+											if (["sse_v2", "sse_v2.1"].includes(protocol)) {
+												close_stream();
+											}
 										}
 									};
 									if (event_id in pending_stream_messages) {
@@ -1048,7 +1053,10 @@ export function api_factory(
 							)
 						);
 					} else if (event_callbacks[event_id]) {
-						if (_data.msg === "process_completed") {
+						if (
+							_data.msg === "process_completed" &&
+							["sse", "sse_v1", "sse_v2", "sse_v2.1"].includes(config.protocol)
+						) {
 							unclosed_events.delete(event_id);
 							if (unclosed_events.size === 0) {
 								close_stream();
@@ -1061,6 +1069,9 @@ export function api_factory(
 							pending_stream_messages[event_id] = [];
 						}
 						pending_stream_messages[event_id].push(_data);
+					}
+					if (_data.msg === "close_stream") {
+						close_stream();
 					}
 				};
 				event_stream.onerror = async function (event) {