+++ js/file/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/file
 
+## 0.5.7
+
+### Dependency updates
+
+- @gradio/client@0.14.0
+- @gradio/upload@0.8.0
+- @gradio/wasm@0.9.0
+
 ## 0.5.6
 
 ### Dependency updates